+++ lib/eventum/class.db_helper.php
@@ -16,10 +16,6 @@ use Eventum\Db\Adapter\NullAdapter;
 use Eventum\Db\DatabaseException;
 use Eventum\Monolog\Logger;
 
-/**
- * Class to manage all tasks related to the DB abstraction module. This is only
- * useful to maintain a data dictionary of the current database schema tables.
- */
 class DB_Helper
 {
     const DEFAULT_ADAPTER = 'PdoAdapter';
@@ -78,7 +74,6 @@ class DB_Helper
 
     /**
      * Get database config.
-     * load it from setup, fall back to legacy config.php constants
      *
      * @return array
      */