+++ lib/eventum/class.user.php
@@ -69,6 +69,7 @@ class User
      *
      * @param   int $customer_contact_id The customer contact ID
      * @return  int The user ID
+     * @deprecated use UserRepository
      */
     public static function getUserIDByContactID($customer_contact_id)
     {
@@ -94,6 +95,7 @@ class User
      *
      * @param   int $customer_contact_id The customer contact ID
      * @return  string The user's email address
+     * @deprecated use UserRepository
      */
     public static function getEmailByContactID($customer_contact_id)
     {